+++ lib/mini_kraken/glue/run_star_expression.rb
@@ -9,19 +9,15 @@ module MiniKraken
     class RunStarExpression
       attr_reader :env
 
-      # @param var_name [String]
-      # @param goal [Core::Goal]
-      def initialize(var_name, goal)
-        @env = FreshEnv.new([var_name], goal)
-      end
-
-      def var
-        env.vars.values.first
+      # @param var_names [String, Array<String>] One variable name or an array of names
+      # @param goal [Core::Goal, Array<Core::Goal>] A single goal or an array of goals to conjunct
+      def initialize(var_names, goal)
+        vnames = var_names.kind_of?(String) ? [var_names] : var_names
+        @env = FreshEnv.new(vnames, goal)
       end
 
       def run
-        result = nil
-        next_result = nil
+        result = []
         solver = env.goal.attain(env)
         # require 'debug'
         loop do
@@ -30,24 +26,44 @@ module MiniKraken
           outcome = solver.resume
           break if outcome.nil?
 
-          if result # ... more than one result...
-            if outcome.successful?
-              next_result.append(Core::ConsCell.new(var.quote(outcome)))
-            else
-              next_result.append(Core::NullList)
-            end
-            next_result = next_result.cdr
-          elsif outcome.successful?
-            env.propagate(outcome)
-            result = Core::ConsCell.new(var.quote(outcome))
-            next_result = result
-          else
-            result = Core::NullList
-            next_result = result
-          end
+          env.propagate(outcome) if result.empty? && outcome.successful?
+          result << build_solution(outcome)
         end
 
-        result
+        format_solutions(result)
+      end
+
+      private
+
+      # @return [Array] A vector of assignment for each variable
+      def build_solution(outcome)
+        sol = env.vars.values.map do |var|
+          outcome.successful? ? var.quote(outcome) : nil
+        end
+
+        sol
+      end
+
+      # Transform the solutions into sequence of conscells.
+      # @param solutions [Array<Array>] An array of solution.
+      # A solution is in itself an array of bindings (one per variable)
+      def format_solutions(solutions)
+        solutions_as_list = solutions.map { |sol| arr2list(sol, true) }
+        arr2list(solutions_as_list, false)
+      end
+
+      # Utility method. Transform an array into a ConsCell-based list.
+      # @param anArray [Array]
+      # @param simplify [Boolean]
+      def arr2list(anArray, simplify)
+        return anArray[0] if anArray.size == 1 && simplify
+
+        new_tail = nil
+        anArray.reverse_each do |elem|
+          new_tail = Core::ConsCell.new(elem, new_tail)
+        end
+
+        new_tail
       end
     end # class
   end # module